refactor(k8s): remove obsolete resources; configure network in terraform
This commit removes several Kubernetes resources that are no longer needed and adds a new networking configuration instead. This update was done as a part of improving the system's efficiency and to better manage the network traffic. The new network configuration now includes a proxy server setup. Ansible playbook is also updated accordingly to use this new configuration. The secrets related to the proxy server are securely updated in the secrets.yml file.